refactor: 룰렛 컴포넌트 리팩토링
