+++ src/components/Roulette/index.tsx
@@ -1,6 +1,6 @@
-import { useEffect, useRef, useCallback } from 'react';
-import { drawItemBackground, drawItemText, findResultItem } from './helpers';
-import type { ItemInfo, Circle } from './types';
+import { useEffect, useState, useRef, useCallback } from 'react';
+import { findRandomRatio, findItemByRatio } from './helpers';
+import type { ItemInfo, Circle, Ratio } from './types';
 import styled from '@emotion/styled';
 
 const ROTATE_DURATION = 2000; // 회전 시간
@@ -13,7 +13,78 @@ interface Props {
 
 const Roulette = ({ items, colors }: Props) => {
   const canvasRef = useRef<HTMLCanvasElement>(null);
+  const [isRotating, setIsRotating] = useState(false);
 
+  /**
+   * 룰렛에서 아이템의 배경을 그립니다.
+   */
+  const drawItemBackground = useCallback(
+    ({
+      ctx,
+      circle,
+      backgroundColor,
+      startAngle,
+      endAngle
+    }: {
+      ctx: CanvasRenderingContext2D;
+      circle: Circle;
+      backgroundColor: string;
+      startAngle: number;
+      endAngle: number;
+    }) => {
+      const { centerX, centerY, radius } = circle;
+
+      ctx.beginPath();
+      ctx.fillStyle = backgroundColor;
+      ctx.lineWidth = 2;
+      ctx.moveTo(centerX, centerY);
+      ctx.arc(centerX, centerY, radius - 5, startAngle, endAngle);
+      ctx.fill();
+      ctx.stroke();
+      ctx.closePath();
+    },
+    []
+  );
+
+  /**
+   * 룰렛에서 아이템의 텍스트를 그립니다.
+   */
+  const drawItemText = useCallback(
+    ({
+      ctx,
+      circle,
+      radian,
+      startAngle,
+      text
+    }: {
+      ctx: CanvasRenderingContext2D;
+      circle: Circle;
+      radian: number;
+      startAngle: number;
+      text: string;
+    }) => {
+      const { centerX, centerY, radius } = circle;
+
+      const textAngle = startAngle + radian / 2;
+      const textX = centerX + (radius / 2) * Math.cos(textAngle);
+      const textY = centerY + (radius / 2) * Math.sin(textAngle);
+
+      ctx.save(); // 현재 캔버스 상태 저장
+      ctx.translate(textX, textY); // 캔버스의 원점을 텍스트 위치로 이동
+      ctx.rotate(textAngle); // 텍스트를 대각선으로 회전
+
+      ctx.font = `bold 16px Raleway`;
+      ctx.fillStyle = 'black';
+      ctx.textAlign = 'center';
+      ctx.fillText(text, 0, 0); // 회전한 캔버스에 텍스트를 그림
+      ctx.restore();
+    },
+    []
+  );
+
+  /**
+   * 룰렛에서 모든 아이템을 그립니다.
+   */
   const drawItems = useCallback(
     (ctx: CanvasRenderingContext2D, circle: Circle) => {
       const totalRatio = items.reduce((acc, { checked, ratio }) => (!checked ? acc : acc + ratio), 0);
@@ -28,7 +99,6 @@ const Roulette = ({ items, colors }: Props) => {
 
         if (!checked) return;
 
-        // 룰렛 아이템 배경 그리기
         drawItemBackground({
           ctx,
           circle,
@@ -37,7 +107,6 @@ const Roulette = ({ items, colors }: Props) => {
           endAngle
         });
 
-        // 룰렛 아이템 텍스트 그리기
         drawItemText({
           ctx,
           circle,
@@ -49,8 +118,41 @@ const Roulette = ({ items, colors }: Props) => {
         startAngle = endAngle;
       });
     },
-    [items, colors]
+    [items, colors, drawItemBackground, drawItemText]
   );
+
+  /**
+   * 룰렛을 회전시키는 애니메이션을 실행합니다.
+   */
+  const animateRotation = useCallback((canvas: HTMLCanvasElement, ratio: Ratio) => {
+    const { random, total } = ratio;
+
+    canvas.style.transform = `initial`;
+    canvas.style.transition = `initial`;
+
+    setTimeout(() => {
+      const arc = 360 / total;
+      const rotate = random * arc + 360 * ROTATE_COUNT + arc / 2;
+
+      canvas.style.transform = `rotate(-${rotate}deg)`;
+      canvas.style.transition = `${ROTATE_DURATION}ms`;
+    }, 0);
+  }, []);
+
+  const handleClick = useCallback(() => {
+    if (!canvasRef.current) return;
+    if (isRotating) return;
+
+    setIsRotating(true);
+    const ratio = findRandomRatio(items);
+    const result = findItemByRatio(items, ratio.random);
+    animateRotation(canvasRef.current, ratio);
+
+    setTimeout(() => {
+      setIsRotating(false);
+      console.log(result);
+    }, ROTATE_DURATION);
+  }, [animateRotation, isRotating, items]);
 
   useEffect(() => {
     const ctx = canvasRef?.current?.getContext('2d');
@@ -59,13 +161,14 @@ const Roulette = ({ items, colors }: Props) => {
     if (!canvasRef.current.getContext) return;
     if (!ctx) return;
 
-    ctx.clearRect(0, 0, canvasRef.current.width, canvasRef.current.height);
-
     const circle: Circle = {
       centerX: canvasRef.current.width / 2,
       centerY: canvasRef.current.height / 2,
       radius: canvasRef.current.width / 2 - 10
     };
+
+    // 룰렛 초기화
+    ctx.clearRect(0, 0, canvasRef.current.width, canvasRef.current.height);
 
     // 룰렛 배경 영역 그리기
     ctx.fillStyle = '#DBDBDB60';
@@ -77,36 +180,6 @@ const Roulette = ({ items, colors }: Props) => {
     drawItems(ctx, circle);
   }, [items, colors, drawItems]);
 
-  const rotateRoulette = useCallback(
-    (canvas: HTMLCanvasElement) => {
-      const totalRatio = items.reduce((acc, { checked, ratio }) => (!checked ? acc : acc + ratio), 0);
-      const randomValue = Math.floor(Math.random() * totalRatio);
-
-      canvas.style.transform = `initial`;
-      canvas.style.transition = `initial`;
-
-      setTimeout(() => {
-        const arc = 360 / totalRatio;
-        const rotate = randomValue * arc + 360 * ROTATE_COUNT + arc / 2;
-
-        canvas.style.transform = `rotate(-${rotate}deg)`;
-        canvas.style.transition = `${ROTATE_DURATION}ms`;
-      }, 0);
-
-      return findResultItem(items, randomValue);
-    },
-    [items]
-  );
-
-  const handleClick = useCallback(() => {
-    if (!canvasRef.current) return;
-    const result = rotateRoulette(canvasRef.current);
-
-    setTimeout(() => {
-      console.log(result);
-    }, ROTATE_DURATION);
-  }, [rotateRoulette]);
-
   return (
     <RouletteContainer>
       <CanvasContainer>
@@ -115,7 +188,11 @@ const Roulette = ({ items, colors }: Props) => {
           height={500}
           ref={canvasRef}
         />
-        <StartButton onClick={handleClick}>START</StartButton>
+        <StartButton
+          disabled={isRotating}
+          onClick={handleClick}>
+          START
+        </StartButton>
       </CanvasContainer>
       <Cursor
         src="/rouletteCursor.svg"
@@ -154,7 +231,7 @@ const StartButton = styled.button`
   font-weight: 600;
   font-size: 1rem;
 
-  &:active {
+  &:enabled:active {
     margin-left: 3px;
     margin-top: 3px;
     box-shadow: none;
